==> 3. example.Rmd <==
```{r warning=FALSE}

### Load packages and set parameters ###

packages <-
  c(
    "simstudy",
    "gtools",
    "tictoc",
    "mvtnorm",
    "ICS",
    "matlib",
    "ggplot2",
    "dplyr",
    "tidyr",
    "tidyverse",
    "reshape2",
    "patchwork"
  )
invisible(lapply(packages, library, character.only = TRUE))
options(scipen = 999)
options(digits = 7)

set.seed(888888)

n              <- 1000L
num_likert     <- 5L
wc_correlation <- 0.75


seeds          <- as.integer(round(1e8 * runif(n), 0))

tic("Example outputs obtained succesfully: ")

```

```{r}

# Example 34
num_constructs <- 4L
construct_size <- 1L
num_items      <- construct_size * num_constructs

# Correlation matrix
construct_pairs     <- combinations(
  n = num_constructs,
  r = 2,
  repeats.allowed = F,
  v = 1:num_constructs
)
# Manual input
bc_indicator        <- list(c(1, 3), c(2, 4))
bc_correlation      <- 0.5
bc_correlation_sign <- c()
for (pair in seq_len(nrow(construct_pairs))) {
  if (any(sapply(bc_indicator,
                 identical,
                 as.numeric(construct_pairs[pair, ])))) {
    bc_correlation_sign <- append(bc_correlation_sign, bc_correlation)
  }
  else {
    bc_correlation_sign <-
      append(bc_correlation_sign, -1 * bc_correlation)
  }
}

Rho <- matrix(0.0, nrow = num_items, ncol = num_items)
nam <- rep(NA_character_, num_items)
for (i in 0:(num_constructs - 1)) {
  interval                <- 1:construct_size + i * construct_size
  Rho[interval, interval] <- i + 2
  nam[interval]           <- paste0(i + 1)
}
for (block in seq_along(bc_correlation_sign)) {
  construct_1                <- construct_pairs[block, 1]
  construct_2                <- construct_pairs[block, 2]
  first_idx                  <-
    ((construct_1 - 1) * construct_size + 1):(construct_1 * construct_size)
  second_idx                 <-
    ((construct_2 - 1) * construct_size + 1):(construct_2 * construct_size)
  Rho[first_idx, second_idx] <- bc_correlation_sign[block]
  Rho[second_idx, first_idx] <- bc_correlation_sign[block]
}
diag(Rho)                             <- 1.0
Rho[Rho %in% 2:(num_constructs + 1L)] <- wc_correlation

# Answer distributions
dist_sign                    <- rep(1, num_constructs)
dist_sign[bc_indicator[[2]]] <- -1
agree_type                   <- c(0.01, 0.04, 0.15, 0.40, 0.40)
disagree_type                <- rev(agree_type)

baseprobs_agree              <- list()
for (sign in dist_sign) {
  if (sign == 1) {
    baseprobs_agree <-
      append(baseprobs_agree, list(t(
        replicate(construct_size, agree_type)
      )))
  } else {
    baseprobs_agree <-
      append(baseprobs_agree, list(t(
        replicate(construct_size, disagree_type)
      )))
  }
}

baseprobs_agree    <- Reduce(rbind, baseprobs_agree)
baseprobs_disagree <- baseprobs_agree[, ncol(baseprobs_agree):1]
baseprobs          <- list(baseprobs_agree, baseprobs_disagree)

# Generate data set
set.seed(888888)
n_dist   <-
  sample(x = 1:length(baseprobs),
         size = n,
         replace = TRUE)
n_dist    <- table(n_dist)
dX        <- list()
for (dist in seq_along(baseprobs)) {
  dT      <- simstudy::genData(n_dist[dist])
  dX_dist <-
    simstudy::genOrdCat(dT,
                        baseprobs = baseprobs[[dist]],
                        prefix = "q",
                        corMatrix = Rho)
  dX_dist <- as.matrix(dX_dist[,-1])
  dX_dist <- matrix(as.integer(dX_dist), nrow = n_dist[dist])
  dX      <- append(dX, list(dX_dist))
}
dX        <- Reduce(rbind, dX)
dX        <- dX[sample(1:nrow(dX)),]

# Introduce CR
for (cr_pos in 3:4) {
  set.seed(888888)
  e                <- 0.05
  num_cr           <- floor(e * n)
  careless_idx     <-
    sample(x = seq_len(n),
           size = num_cr,
           replace = FALSE)
  X_clean <- dX
  for (i in seq_along(careless_idx)) {
    set.seed(seeds[i])
    row                    <- careless_idx[i]
    ifelse(X_clean[row, cr_pos] > 3, cr_fill <- 1, cr_fill <- 5)
    X_clean[row, cr_pos] <- cr_fill
  }
  X <- X_clean
  
  # ICS
  S_1         <- cov(X)
  S_2         <- cov4(X)
  ratio       <- inv(S_1) %*% S_2
  eigen       <- eigen(ratio, symmetric = FALSE)
  D           <- diag(eigen$values)
  B           <- t(eigen$vectors) # B has rowwise eigenvectors
  m_1         <- colMeans(X)
  m_1         <- t(matrix(replicate(nrow(X), m_1), ncol = nrow(X)))
  X_centered  <- as.matrix(X - m_1)
  Z           <- X_centered %*% t(B)
  Z           <- as.data.frame(Z)
  Z           <- Z[, 1:(ncol(Z)-1)]
  colnames(Z)    <- gsub(x = colnames(Z), pattern = "V", replacement = "IC")
  Z$index <- 1:nrow(Z)
  Z_melt <- melt(Z, "index")
  color_label <- rep("Regular", n)
  for (i in 1:n) {
    if (i %in% careless_idx) {
      color_label[i] <- "CR"
    }
  }
  colnames(Z_melt)[ncol(Z_melt)] <- "score"
  
  p1 <- ggplot(Z_melt, aes(x = index, y = score)) +
    geom_point(aes(colour = rep(color_label, (num_items - 1)))) +
    facet_grid(. ~ variable) +
    scale_color_manual(
      name = "Legend",
      breaks = c("Regular", "CR"),
      values = c("Regular" = "black", "CR" = "red")
    ) +
    theme(
      axis.title.x = element_blank(),
      axis.text.x = element_blank(),
      axis.ticks.x = element_blank(),
      axis.title.y = element_blank()
    )
  
  
  eigenvalue_indicator <- list()
  for (i in 1:num_items) {
    eigenvalue_indicator[i] <- paste0(round(eigen$values, 2)[i])
  }
  heat_map_data <- B[, ncol(B):1]
  names(heat_map_data) <- paste("X", 1:nrow(B))
  heat_map_data_2 <-
    heat_map_data %>%
    as_tibble() %>%
    rownames_to_column("Var1") %>%
    pivot_longer(-Var1, names_to = "Var2", values_to = "Legend") %>%
    mutate(Var1 = factor(Var1, levels = 1:nrow(B)),
           Var2 = factor(gsub("V", "", Var2), levels = 1:nrow(B)))
  heat_map_data_2 <- heat_map_data_2[1:(nrow(heat_map_data_2)-num_constructs), ]
  p2 <- ggplot(heat_map_data_2, aes(Var1, Var2)) +
    geom_tile(aes(fill = Legend)) +
    geom_text(aes(label = round(Legend, 2))) +
    scale_fill_gradient2(
      low = "red3",
      mid = "white",
      high = "blue3",
      limits = c(-1, 1)
    ) +
    guides(fill = guide_colourbar(
      barwidth = 0.5,
      barheight = 5,
      ticks = FALSE
    )) +
    scale_x_discrete(name = "Eigenvector + value", labels = eigenvalue_indicator) +
    scale_y_discrete(name = "", labels = nrow(B):1) +
    theme(plot.title = element_text(hjust = 0.5),
          panel.background = element_blank())
  print(p1 / p2)
  
  file_name <- paste0("3. Plots/example_", cr_pos, ".pdf")
  ggsave(file_name,
         width = 7,
         height = 3,
         dpi = 1200)
}

```

```{r}

# Example 5
num_constructs <- 4L
construct_size <- 3L
num_items      <- construct_size * num_constructs

# Correlation matrix
construct_pairs     <- combinations(
  n = num_constructs,
  r = 2,
  repeats.allowed = F,
  v = 1:num_constructs
)
# Manual input
bc_indicator        <- list(c(1, 3), c(2, 4))
bc_correlation      <- 0.5
bc_correlation_sign <- c()
for (pair in seq_len(nrow(construct_pairs))) {
  if (any(sapply(bc_indicator,
                 identical,
                 as.numeric(construct_pairs[pair, ])))) {
    bc_correlation_sign <- append(bc_correlation_sign, bc_correlation)
  }
  else {
    bc_correlation_sign <-
      append(bc_correlation_sign,-1 * bc_correlation)
  }
}

Rho <- matrix(0.0, nrow = num_items, ncol = num_items)
nam <- rep(NA_character_, num_items)
for (i in 0:(num_constructs - 1)) {
  interval                <- 1:construct_size + i * construct_size
  Rho[interval, interval] <- i + 2
  nam[interval]           <- paste0(i + 1)
}
for (block in seq_along(bc_correlation_sign)) {
  construct_1                <- construct_pairs[block, 1]
  construct_2                <- construct_pairs[block, 2]
  first_idx                  <-
    ((construct_1 - 1) * construct_size + 1):(construct_1 * construct_size)
  second_idx                 <-
    ((construct_2 - 1) * construct_size + 1):(construct_2 * construct_size)
  Rho[first_idx, second_idx] <- bc_correlation_sign[block]
  Rho[second_idx, first_idx] <- bc_correlation_sign[block]
}
diag(Rho)                             <- 1.0  
Rho[Rho %in% 2:(num_constructs + 1L)] <- wc_correlation

# Answer distributions
dist_sign                    <- rep(1, num_constructs)
dist_sign[bc_indicator[[2]]] <- -1
agree_type                   <- c(0.01, 0.04, 0.15, 0.40, 0.40)
disagree_type                <- rev(agree_type)

baseprobs_agree              <- list()
for (sign in dist_sign) {
  if (sign == 1) {
    baseprobs_agree <-
      append(baseprobs_agree, list(t(
        replicate(construct_size, agree_type)
      )))
  } else {
    baseprobs_agree <-
      append(baseprobs_agree, list(t(
        replicate(construct_size, disagree_type)
      )))
  }
}

baseprobs_agree    <- Reduce(rbind, baseprobs_agree)
baseprobs_disagree <- baseprobs_agree[, ncol(baseprobs_agree):1]
baseprobs          <- list(baseprobs_agree, baseprobs_disagree)

# Generate data set
n_dist   <-
  sample(x = 1:length(baseprobs),
         size = n,
         replace = TRUE)
n_dist    <- table(n_dist)
dX        <- list()
for (dist in seq_along(baseprobs)) {
  dT      <- simstudy::genData(n_dist[dist])
  dX_dist <-
    simstudy::genOrdCat(dT,
                        baseprobs = baseprobs[[dist]],
                        prefix = "q",
                        corMatrix = Rho)
  dX_dist <- as.matrix(dX_dist[,-1])
  dX_dist <- matrix(as.integer(dX_dist), nrow = n_dist[dist])
  dX      <- append(dX, list(dX_dist))
}
dX        <- Reduce(rbind, dX)
dX        <- dX[sample(1:nrow(dX)),]

# Introduce CR
set.seed(16)
e                <- 0.05
num_cr           <- floor(e * n)
careless_idx     <-
  sample(x = seq_len(n),
         size = num_cr,
         replace = FALSE)
cr_array         <- 10:num_items
X_clean <- dX
for (i in seq_along(careless_idx)) {
  set.seed(seeds[i])
  row                    <- careless_idx[i]
  X_clean[row, cr_array] <-
    sample.int(5, size = length(cr_array), replace = TRUE)
}
X <- X_clean

# ICS
S_1         <- cov(X)
S_2         <- cov4(X)
ratio       <- inv(S_1) %*% S_2
eigen       <- eigen(ratio, symmetric = FALSE)
D           <- diag(eigen$values)
B           <- t(eigen$vectors) # B has rowwise eigenvectors
m_1         <- colMeans(X)
m_1         <- t(matrix(replicate(nrow(X), m_1), ncol = nrow(X)))
X_centered  <- as.matrix(X - m_1)
Z           <- X_centered %*% t(B)
Z           <- as.data.frame(Z)
Z           <- Z[, 1:(ncol(Z)-1)]
colnames(Z) <- gsub(x = colnames(Z), pattern = "V", replacement = "IC")
Z$index <- 1:nrow(Z)
Z_melt <- melt(Z, "index")
color_label <- rep("Regular", n)
for (i in 1:n) {
  if (i %in% careless_idx) {
    color_label[i] <- "CR"
  }
}
colnames(Z_melt)[ncol(Z_melt)] <- "score"

p1 <- ggplot(Z_melt, aes(x = index, y = score)) +
  geom_point(aes(colour = rep(color_label, (num_items-1)))) +
  facet_grid(. ~ variable) +
  scale_color_manual(
    name = "Legend",
    breaks = c("Regular", "CR"),
    values = c("Regular" = "black", "CR" = "red")
  ) +
  theme(
    axis.title.x = element_blank(),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    axis.title.y = element_blank()
  )


eigenvalue_indicator <- list()
for (i in 1:num_items) {
  eigenvalue_indicator[i] <- paste0(round(eigen$values, 2)[i])
}
heat_map_data <- B[, ncol(B):1]
names(heat_map_data) <- paste("X", 1:nrow(B))
heat_map_data_2 <-
  heat_map_data %>%
  as_tibble() %>%
  rownames_to_column("Var1") %>%
  pivot_longer(-Var1, names_to = "Var2", values_to = "Legend") %>%
  mutate(Var1 = factor(Var1, levels = 1:nrow(B)),
         Var2 = factor(gsub("V", "", Var2), levels = 1:nrow(B)))
heat_map_data_2 <- heat_map_data_2[1:(nrow(heat_map_data_2) - num_items), ]
p2 <- ggplot(heat_map_data_2, aes(Var1, Var2)) +
  geom_tile(aes(fill = Legend)) +
  geom_text(aes(label = round(Legend, 2))) +
  scale_fill_gradient2(
    low = "red3",
    mid = "white",
    high = "blue3",
    limits = c(-1, 1)
  ) +
  guides(fill = guide_colourbar(
    barwidth = 0.5,
    barheight = 5,
    ticks = FALSE
  )) +
  scale_x_discrete(name = "Eigenvector + value", labels = eigenvalue_indicator) +
  scale_y_discrete(name = "", labels = nrow(B):1) +
  theme(plot.title = element_text(hjust = 0.5),
        panel.background = element_blank())
print(p1 / p2)

file_name <- paste0("3. Plots/example_5.pdf")
ggsave(file_name,
       width = 11,
       height = 7,
       dpi = 1200)



toc()

```
